final-project/kafka-consumer/requirements.txt
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
peewee==3.17.3
psycopg2-binary==2.9.9
